+++ backend/api/models/games_models.py
@@ -37,6 +37,7 @@ class Games(MediaMixin, db.Model):
     hltb_total_complete_time = db.Column(db.String(20))
     api_id = db.Column(db.Integer, nullable=False)
     lock_status = db.Column(db.Boolean, default=1)
+    last_api_update = db.Column(db.DateTime)
 
     genres = db.relationship("GamesGenre", backref="games", lazy=True)
     platforms_rl = db.relationship("GamesPlatforms", backref="games", lazy=True)