+++ frontend/src/components/comingNext/nextMedia.jsx
@@ -1,18 +1,20 @@
-import {zeroPad} from "@/lib/utils";
+import {cn, zeroPad} from "@/lib/utils";
 import {MediaCard} from "@/components/reused/MediaCard";
 
 
 export const NextMedia = ({ mediaType, media }) => {
+    const position = mediaType === ("movies" || "games") ? "bottom-[0px]" : "bottom-[32px]";
+
     return (
         <MediaCard media={media} mediaType={mediaType}>
             {(mediaType === "anime" || mediaType === "series") &&
-                <div className="z-20 absolute flex justify-center items-center bottom-[0px] h-[32px] w-full opacity-90
-                border border-black bg-neutral-950">
+                <div className={cn("absolute flex justify-center items-center h-[32px] w-full opacity-95 " +
+                    "bg-neutral-900 border border-black border-solid", position)}>
                     S{zeroPad(media.season_to_air)}&nbsp;-&nbsp;E{zeroPad(media.episode_to_air)}
                 </div>
             }
-            <div className="z-20 absolute h-[32px] flex items-center justify-center w-full opacity-90 bg-teal-950
-            border border-x-black border-b-black border-t-transparent rounded-bl-md rounded-br-md">
+            <div className="flex justify-center items-center h-[32px] w-full opacity-90 bg-gray-900 border
+                border-x-black border-b-black rounded-bl-md rounded-br border-t-transparent">
                 {media.date}
             </div>
         </MediaCard>