+++ frontend/src/components/media/general/RefreshMedia.jsx
@@ -1,13 +1,14 @@
 import {toast} from "sonner";
 import {FaRedo} from "react-icons/fa";
+import {createLocalDate} from "@/lib/utils";
 import {useLoading} from "@/hooks/LoadingHook";
 import {Tooltip} from "@/components/ui/tooltip";
 import {LoadingIcon} from "@/components/primitives/LoadingIcon";
-import {createLocalDate} from "@/lib/utils.jsx";
 
 
-export const RefreshMedia = ({ updateRefresh, mutateData, lastUpdate }) => {
+export const RefreshMedia = ({ updateRefresh, mutateData, lastApiUpdate }) => {
     const [isLoading, handleLoading] = useLoading();
+    const lastRefresh = lastApiUpdate ? createLocalDate(lastApiUpdate) : "Never";
 
     const handleRefresh = async () => {
         const response = await handleLoading(updateRefresh);
@@ -22,7 +23,7 @@ export const RefreshMedia = ({ updateRefresh, mutateData, lastUpdate }) => {
             {isLoading ?
                 <LoadingIcon size={8} cssOverride={{marginTop: 12}}/>
                 :
-                <Tooltip text="Refresh metadata" subText={`Last refresh: ${createLocalDate(lastUpdate)}`} side="left">
+                <Tooltip text="Refresh metadata" subText={`Last refresh: ${lastRefresh}`} side="left">
                     <div role="button" onClick={handleRefresh}>
                         <FaRedo size={18} className="mt-2"/>
                     </div>